Move describe workflow to api package
